WinUIBackend: Support displaying multiple alerts at once
